+++ Services/CompanyService/CompanyService.DAL/Сontracts/CreateVacancy.cs
@@ -8,7 +8,7 @@ public record
         string post,
         int min_salary,
         int? max_salary,
-        ulong? education_id,
+        int? education_id,
         short? experience,
         string? description,
         WorkFormat? work_format,
